time: dont round up timespec functions
using DIV_ROUND_CLOSEST will round up the second value of timespec when the
nanosecond digits are greater than half a second. This will result in extra
time added to the timespec result and confuse timing functions like
gettimeofday.

for example a value of 15779166000 nanoseconds would be convert to
	tv_sec = 16; tv_nsec = 779166000.
but should be:
	tv_sec = 15; tv_nsec = 779166000.

Instead of DIV_ROUND_CLOSEST use plain division.

Signed-off-by: Axel Haslam <ahaslam@baylibre.com>