time: dont round up timespec functions
this reverts commit 2169fff4c9f8bd81b438214c6a3537e68bee02b2

using DIV_ROUND_CLOSEST will round up the seconds value of
timespec when the nanosecond digits are ammount for more
than half a second. This will add an extra second and produce
wrong results from timing functions like gettimeofday.

for example a value of 15779166000 nanoseconds would be converted to:
    tv_sec = 16; tv_nsec = 779166000.
but should be:
    tv_sec = 15; tv_nsec = 779166000.

Instead of DIV_ROUND_CLOSEST use plain division.

Testing Done: In a dummy app, call gettimeofday in a loop and verify
              that time always increases.
Bug Number: SW-2202

Signed-off-by: Axel Haslam <ahaslam@baylibre.com>